backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.4

# Speech-to-text
openai-whisper==20231106
SpeechRecognition==3.10.0

# Audio processing
librosa==0.10.0
soundfile==0.12.1
PyAudio==0.2.13

# Speaker diarization (optional - can use pyannote.audio)
pyannote.audio==3.0.1

# LLM integration
ollama==0.1.37

# Logging and monitoring
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS support
python-cors==1.0.1

